requirements.txt
numpy==1.21.3
Pillow==8.4.0
PyGLM==2.5.4
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
PySide6==6.2.1
shiboken6==6.2.1
